+++ packages/keyring-eth-ledger-bridge/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Bump `@metamask/keyring-api` from `^23.3.0` to `^23.4.0` ([#583](https://github.com/MetaMask/accounts/pull/583))
+
 ## [12.2.0]
 
 ### Added